+++ InvestmentBuilderLib/InvestmentData.cs
@@ -20,7 +20,7 @@ namespace InvestmentBuilder
         double Price { get; }
         InvestmentInformation CompanyData { get; }
         void UpdateRow(DateTime valuationDate, DateTime previousDate);
-        void ChangeShareHolding(DateTime valuationDate, int quantity);
+        void ChangeShareHolding(DateTime valuationDate, double quantity);
         void AddNewShares(DateTime valuationDate, Stock stock);
         void UpdateClosingPrice(DateTime valuationDate, double dClosing);
         void UpdateDividend(DateTime valuationDate, double dDividend);
@@ -49,7 +49,7 @@ namespace InvestmentBuilder
             _investmentRecordData.RollInvestment(_userToken, Name, valuationDate, previousDate);
         }
 
-        public void ChangeShareHolding(DateTime valuationDate, int quantity)
+        public void ChangeShareHolding(DateTime valuationDate, double quantity)
         {
             _investmentRecordData.UpdateInvestmentQuantity(_userToken, Name, valuationDate, quantity);
         }
